+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v0.2.0
 _2021-05-11_
 
 * Add support for Markdown (thanks to @bastikempken)
+* Adjust to new plugin API
 * Housekeeping